JobWidget now creates URL with RecruiterId instead of RecruiterUrl.
Summary: Depends on D1745. Changed the naming for 'r' to 'rid' and changed the json extracted 'nameUrl' to 'id'. This causes the widget URL that is created to spit out 'rid=123' instead of the former 'r=uitzendbureau-nl'

Test Plan:
**Make sure to clone the 'for-developer' repo. This is the devhub project.**
Phabricator also says: Repository rFORDEVELOPERS For Developers

Manually edit 'javascripts/api-clients/uitzendbureau-nl-api.js' and change #Line40
> this.rootUrl = 'http://www.uitzendbureau.nl/api';
to
> this.rootUrl = 'http://uzb.xxx.uzbdev.nl/api';

Jekyll serve the pages locally: (Go to project root, have Jekyll installed, type
> jekyll serve
and run the website locally.)

Go to localhost:4000, click `Widgets` and click `Vacaturewidget`
Create a widget URL with uitzendbureau-nl as the recruiter.
This should give you "rid=953" in the url instead of "r=uitzendbureau-nl"

Reviewers: janek, #uitzendbureau.nl, dennis

Reviewed By: janek, #uitzendbureau.nl, dennis

Differential Revision: http://phabricator.uzbdev.nl/D1805