Refactoring the API documentation. Preventing confusion about the use of URL encoded POST method.
Summary:
Added some extra info to make it even more clear that the developer should use url-encoded POST data to submit a request to the API.
Added the type of POST method information everywhere we specify the data should be submitted with a POST. Also changed some text that described 'the data' and changed it to 'the response' to clearly specify the objects.

Test Plan:
Check out the changes, there are only text changes.

Extra info for @bas because maybe it's the first time using Jekyll:

[[ https://jekyllrb.com/docs/installation/ | Install Jekyll ]]

How to compile / run / serve Jekyll pages locally:
Go to the project root.
Type this in your terminal `jekyll serve`
The webpage is now being built and served on `localhost:4000`
Check out the page locally.
When the diff is accepted by you (and the next reviewer) the changes are pushed to Github.
Then Github will serve the github repo from `https://github.com/TeamYellowYellow/for-developers`, we use it at `http://developer.uitzendbureau.nl/`

(Small detail: because this repo is hosted for a company (TeamYellowYellow on Github), github requires the branch to be named `gh-pages`, because the `master` branch can still be used as the company website. So we do not push to master.)

Also, the commit message does not follow convention because this is not the `/uzb` repo. This commit will not be shown in the email when the uzb website is updated, so we don't have `Bugfix`or `refactoring` tags/labels here.

Reviewers: ivo, dennis, #uitzendbureau.nl

Reviewed By: ivo, dennis, #uitzendbureau.nl

Subscribers: bas

Differential Revision: http://phabricator.uzbdev.nl/D1928