Use new package-lock.json to set fixed dev dependencies
Summary: Newer versions of NPM will create this each time. So this just puts it in there, so NPM won't need to create it each time.

Test Plan: Edit `package.json` (for example, change grunt version to `1.0.2`), run `$ npm i` and see that the packages are installed according to `package-lock.json`.

Reviewers: dennis, #uitzendbureau.nl

Reviewed By: dennis, #uitzendbureau.nl

Differential Revision: http://phabricator.hippo.farm/D3560